fix: use exponential back off
This is to handle Google form's data not available when there're many people sending request at the same time. Which will test if data is available or not. If not it will wait 2 seconds and if still not it will double the wait time until data is reachable. And will give up after 10 times of retries (roundly 30 min).

fix #1